feat(wren-ui): refactor strategy a bit (put strategy of data sources into different files) and add pg strategy
