+++ wren-ui/src/apollo/server/factories/dataSourceStrategy.ts
@@ -1,0 +1,17 @@
+import { CompactTable } from '../connectors/connector';
+import { Model, ModelColumn } from '../repositories';
+import { AnalysisRelationInfo } from '../types';
+
+export interface IDataSourceStrategy {
+  saveDataSource(properties: any): Promise<any>;
+  listTable({
+    formatToCompactTable,
+  }: {
+    formatToCompactTable: boolean;
+  }): Promise<CompactTable[] | any[]>;
+  saveModels(models: any): Promise<any>;
+  analysisRelation(
+    models: Model[],
+    columns: ModelColumn[],
+  ): Promise<AnalysisRelationInfo[]>;
+}
